+++ astropy/io/votable/tests/vo_test.py
@@ -611,6 +611,14 @@ class TestParse:
         assert fields[0].name == "int"
         assert fields[0].values.min == -1000
 
+    def test_get_info_by_id(self):
+        info = self.votable.get_info_by_id('QUERY_STATUS')
+        assert info.value == 'OK'
+
+        if self.votable.version != '1.1':
+            info = self.votable.get_info_by_id("ErrorInfo")
+            assert info.value == "One might expect to find some INFO here, too..."
+
 
 class TestThroughTableData(TestParse):
     def setup_class(self):